Use pytest properly
Replaced fixtures with parametrized tests - the more "pytesty" way.
Tests sped up, too.

Tests passed before & after, singly & all together.
