+++ tests/func/test_pass_logging.py
@@ -1,191 +1,64 @@
 # for 'pass' cases, they should be reported to SNS only if VERBOSE is trueish
 
-import boto3
-import os
 import json
 from moto import mock_s3, mock_sns, mock_sqs
 import pytest
+import tests.utils as u
 
 from fx_sig_verify.validate_moz_signature import (lambda_handler, )  # noqa: E402
 
-# Constants
-bucket_name = 'pseudo-bucket'
-sqs_name = "test-queue"
 
-
-class DummyContext(object):
-    aws_request_id = 'DUMMY ID'
-
-
-dummy_context = DummyContext()
-
-
-@pytest.fixture(scope='module', autouse=True)
-def disable_production_filtering():
-    # we want to process "invalid" files during testing
-    os.environ['PRODUCTION'] = "0"
-
-
-@pytest.fixture(scope='module', autouse=True)
-def disable_xray():
-    # at the moment, 'moto' doesn't support xray, so I've hacked fleece to allow
-    # an environment variable to disable them.
-    os.environ['XRAY_DISABLE'] = 'True'
-
-
-def build_event(bucket, key):
-    record = {'s3': {'bucket': {'name': bucket},
-                     'object': {'key': key},
-                     },
-              }
-    return {'Records': [record, ]}
-
-
-@pytest.fixture()
-def good_files():
-    payload = [('32bit.exe', None),  # signed with older valid key
-               ('32bit_new.exe', None),  # signed with current valid key
-               # valid, but tests S3 naming issue (issue #14)
-               ('32bit_new.exe', '32bit new.exe'),
-               ]
-    print(payload)
-    return payload
-
-
-def delete_verbose():
+@mock_s3
+@mock_sns
+@mock_sqs
+@pytest.mark.parametrize('fname', u.good_file_names_list)
+@pytest.mark.parametrize('setter', u.set_verbose_false_list[:1])
+def test_pass_no_message_when_no_verbose(setter, fname):
+    queue = u.setup_aws_mocks()
+    bucket = u.create_bucket()
+    # given that verbose is not set
+    setter()
+    # when a good file is processed
+    bucket_name, key_name = u.upload_file(bucket, fname)
+    event = u.build_event(bucket_name, key_name)
+    response = lambda_handler(event, u.dummy_context)
+    # THEN there should be no message
+    count, msg_json = u.get_one_message(queue)
     try:
-        del os.environ['VERBOSE']
-    except KeyError:
-        pass
+        msg_dict = json.loads(msg_json)
+        msg = msg_dict['Message']
+    except ValueError:
+        msg = ""
 
+    # print things that will be useful to debug
+    print("response:", response)
+    print("message:", msg)
+    print("count:", count)
 
-def zero_verbose():
-    os.environ['VERBOSE'] = '0'
-
-
-def unset_verbose():
-    os.environ['VERBOSE'] = ''
-
-
-@pytest.fixture
-def set_verbose_false():
-    return [delete_verbose, zero_verbose, unset_verbose]
-
-
-def one_verbose():
-    os.environ['VERBOSE'] = '1'
-
-
-def two_verbose():
-    # 2 is debug level
-    os.environ['VERBOSE'] = '2'
-
-
-def true_verbose():
-    os.environ['VERBOSE'] = 'True'
-
-
-@pytest.fixture
-def set_verbose_true():
-    return [one_verbose, two_verbose, true_verbose]
-
-
-def create_bucket():
-    conn = boto3.resource('s3', region_name='us-east-1')
-    bucket = conn.create_bucket(Bucket=bucket_name)
-    return bucket
-
-
-def upload_file(bucket, ftuple):
-    # issue #14 requires a difference to test
-    filename, key_name = ftuple
-    if key_name is None:
-        key_name = filename
-    data_dir = os.path.join(os.path.dirname(__file__), '..', 'data')
-    fname = os.path.join(data_dir, filename)
-    bucket.put_object(Body=open(fname, 'r'), Key=key_name)
-    # S3 helpfully gives us the key with '+', but expects back with ' ', so
-    # mimic that behavior
-    s3name = key_name.replace('+', ' ')
-    return (bucket_name, s3name)
-
-
-def setup_aws_mocks():
-    # mock the SNS topic & pass via environment
-    client = boto3.client("sns")
-    client.create_topic(Name="some-topic")
-    response = client.list_topics()
-    topic_arn = response["Topics"][0]['TopicArn']
-    os.environ['SNSARN'] = topic_arn
-
-    # setup an sqs queue to pull the message from
-    sqs_conn = boto3.resource('sqs', region_name='us-east-1')
-    sqs_conn.create_queue(QueueName=sqs_name)
-
-    client.subscribe(TopicArn=topic_arn,
-                     Protocol="sqs",
-                     Endpoint="arn:aws:sqs:us-east-1:123456789012:test-queue")
-    queue = sqs_conn.get_queue_by_name(QueueName=sqs_name)
-    return queue
-
-
-def get_one_message(queue):
-    # get more than one to detect double push
-    messages = queue.receive_messages(MaxNumberOfMessages=10)
-    return (len(messages), messages[0].body) if len(messages) else (0, '')
+    # actual criteria to pass
+    assert "pass" in response['results'][0]['status']
+    assert count is 0 and msg is ""
 
 
 @mock_s3
 @mock_sns
 @mock_sqs
-def test_pass_no_message_when_no_verbose(set_verbose_false, good_files):
-    queue = setup_aws_mocks()
-    bucket = create_bucket()
-    # Given that VERBOSE is not set
-    for falsey in set_verbose_false:
-        falsey()
-        # WHEN a good file is processed
-        for fname in good_files:
-            bucket_name, key_name = upload_file(bucket, fname)
-            event = build_event(bucket_name, key_name)
-            response = lambda_handler(event, dummy_context)
-            # THEN there should be no message
-            count, msg_json = get_one_message(queue)
-            try:
-                msg_dict = json.loads(msg_json)
-                msg = msg_dict['Message']
-            except ValueError:
-                msg = ""
-
-            # print things that will be useful to debug
-            print("response:", response)
-            print("message:", msg)
-            print("count:", count)
-
-            # actual criteria to pass
-            assert "pass" in response['results'][0]['status']
-            assert count is 0 and msg is ""
-
-
-@mock_s3
-@mock_sns
-@mock_sqs
-def test_pass_message_when_verbose(set_verbose_true, good_files):
-    queue = setup_aws_mocks()
-    bucket = create_bucket()
+@pytest.mark.parametrize('fname', u.good_file_names_list)
+@pytest.mark.parametrize('setter', u.set_verbose_true_list[:1])
+def test_pass_message_when_verbose(setter, fname):
+    queue = u.setup_aws_mocks()
+    bucket = u.create_bucket()
     # Given that VERBOSE is set
-    for truthy in set_verbose_true:
-        truthy()
-        # WHEN a good file is processed
-        for fname in good_files:
-            bucket_name, key_name = upload_file(bucket, fname)
-            event = build_event(bucket_name, key_name)
-            response = lambda_handler(event, dummy_context)
-            print("response:", response)
-            # THEN there should be no message
-            count, msg_json = get_one_message(queue)
-            msg_dict = json.loads(msg_json)
-            msg = msg_dict['Message']
-            print("message:", msg)
-            assert "pass" in response['results'][0]['status']
-            assert count is 1 and msg.startswith('pass for')
+    setter()
+    # WHEN a good file is processed
+    bucket_name, key_name = u.upload_file(bucket, fname)
+    event = u.build_event(bucket_name, key_name)
+    response = lambda_handler(event, u.dummy_context)
+    print("response:", response)
+    # THEN there should be no message
+    count, msg_json = u.get_one_message(queue)
+    msg_dict = json.loads(msg_json)
+    msg = msg_dict['Message']
+    print("message:", msg)
+    assert "pass" in response['results'][0]['status']
+    assert count is 1 and msg.startswith('pass for')
